requirements.txt
# Web Framework
streamlit

# Google Gemini AI Integration
google-genai

# LLM Orchestration & Agent Framework
langchain
langchain-core
langchain-community
langchain-google-genai
langgraph

# Vector Search & Embeddings
faiss-cpu
numpy

# Data Processing
pandas

# Database
psycopg2-binary

# Document Processing
pypdf2
python-docx
openpyxl

# Keyword Search
rank-bm25

# Utilities
sift-stack-py

# Optional: For development
# pytest
# pytest-cov
# black
# flake8
